Changed the way *_memsize functions are implemented.
Instead of having a separate implementation for a *_create function
and a *_memsize function, I now have a *_create_impl function that
has two modes - one for allocation and the other for state
initialization. This requires some boilerplate in *_create_impl
function, but everything is less bugprone this way, since it is
a lot harder to have a mismatch between the allocation and
state initialization code.

This also changes the signature of *_memsize functions. Each *_memsize
function now takes the same template parameters as the corresponding
*_create function and the same parameters, except for the allocated
memory parameter.

All of the preceding WIP commits are parts of this change.