onedim.h: I've replaced the global twiddle
table, stored directly in Fft, with a separate twiddle
table for each step. Pointers to those tables are stored
in Step instances. This should make the code easier to
understand, since it removes the implicit convention about
the layout of the global twiddle table that needed to be
respected by both init_twiddle and the pass implementations.

TODO: Do the same in multidim.h