Got rid of Step, init_steps and
the step related stuff in fft_impl. I'll replace that functionality with
three separate, simpler implementations of Fft state type,
initialization and computation - one for tiny sizes (the ones small
enough that each gets a separate template specialization where the size
is a template parameter), one for small sizes (the ones that are larger
than tiny sizes but are still expected to fit into the cache so that the
recursive algorithm isn't needed) and one for large sizes (the ones
where the recursive algorithm is used). This commit only implements tiny
and small transforms - it temporarily breaks the large transforms.