Added method load_deinterleaved to all the
simd implementation classes and used it in
Complex::load. This allowed me to implement
AvxFloat::load_deinterleaved using vinsertf128 with
a memory operand instead of vperm2f128, as the intel
optimization manual suggests in 11.11.2. This reduces
the pressure on port 5 and makes the algorithm significantly
faster (I've measured a 7% speedup in some cases).